+++ app/Livewire/TransferForm.php
@@ -2,10 +2,13 @@
 
 namespace App\Livewire;
 
+use App\Enums\Role;
 use App\Models\Kardex;
 use App\Models\Stock;
 use App\Models\Transfer;
+use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\DB;
+use Illuminate\Support\Facades\Hash;
 use Livewire\Attributes\On;
 use Livewire\Component;
 
@@ -37,43 +40,45 @@ class TransferForm extends Component
         }
     }
 
-    public function remove(){
+    public function remove($password){
 
-        $last = Kardex::latest()->first();
-        $this->verify = false;
-        foreach ($this->details as $detail){
-            if($detail->kardex_id == $last->id){
-                $this->verify = true;
-                break;
-            }
-        }
-
-        if(!$this->verify){
-            $this->js('Swal.fire({icon: "error", title: "Error", text: "No se puede Eliminar", showConfirmButton: false, timer: 3000})');
-            return;
-        }
-        try {
-            DB::transaction(function () {
-                foreach ($this->details as $detail) {
-                    $stock = Stock::where('product_id', $detail->product_id)
-                        ->where('store_id', $detail->store_id)
-                        ->lockForUpdate()
-                        ->first();
-
-                    $stock->quantity += $detail->quantity - $detail->kardex->quantity;
-                    $stock->save();
-
-                    $k = $detail->kardex->id;
-                    $detail->delete();
-                    Kardex::destroy($k);
+        if(Hash::check($password, Auth::user()->password) && Auth::user()->role == Role::ADMIN) {
+            $last = Kardex::latest()->first();
+            $this->verify = false;
+            foreach ($this->details as $detail) {
+                if ($detail->kardex_id == $last->id) {
+                    $this->verify = true;
+                    break;
                 }
-                Transfer::destroy($this->_id);
-            });
-        } catch (\Throwable $e) {
-            $e->getMessage();
-        }
+            }
 
-        return $this->redirect(route('admin.transfers'));
+            if (!$this->verify) {
+                $this->js('Swal.fire({icon: "error", title: "Error", text: "No se puede Eliminar", showConfirmButton: false, timer: 3000})');
+                return;
+            }
+            try {
+                DB::transaction(function () {
+                    foreach ($this->details as $detail) {
+                        $stock = Stock::where('product_id', $detail->product_id)
+                            ->where('store_id', $detail->store_id)
+                            ->lockForUpdate()
+                            ->first();
+
+                        $stock->quantity += $detail->quantity - $detail->kardex->quantity;
+                        $stock->save();
+
+                        $k = $detail->kardex->id;
+                        $detail->delete();
+                        Kardex::destroy($k);
+                    }
+                    Transfer::destroy($this->_id);
+                });
+            } catch (\Throwable $e) {
+                $e->getMessage();
+            }
+
+            return $this->redirect(route('admin.transfers'));
+        }
     }
 
     public function render()
